Separate model definition and fitting
